macaddrsetup: Correct WLAN MAC address format
The macaddrsetup service writes the WLAN MAC address to a file in /data
so it can be read by the kernel driver later in boot. Unfortunately the
file output by the service doesn't match the format the kernel is
expecting, so the driver silently goes on to make up a MAC for the
interface. This corrects the format of the file to match the kernel.

Change-Id: I88d8524f35b169fc5a04fe1449137ae9f0cd64a5